Finish : philosopher_bonuns

+++ philosopher/philo_bonus/get_status_philosopher.c
@@ -6,7 +6,7 @@
 /*   By: haryu <haryu@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/12 07:13:27 by haryu             #+#    #+#             */
-/*   Updated: 2022/07/12 07:53:41 by haryu            ###   ########.fr       */
+/*   Updated: 2022/07/13 15:41:54 by haryu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,12 +16,9 @@ static void	fork_release(sem_t *fork_sem, int number)
 {
 	int	i;
 
-	i = 0;
-	while (i < number)
-	{
+	i = -1;
+	while (++i < number)
 		sem_post(fork_sem);
-		i++;
-	}
 	return ;
 }
 
@@ -30,16 +27,33 @@ static void	kill_others(pid_t temp, pid_t *philosopher, int max_len)
 	int	i;
 
 	i = -1;
-	(void)temp;
 	while (++i < max_len)
 	{
-		printf("kill %d\n", i);
+		if (temp == *philosopher)
+			continue ;
 		kill(philosopher[i], SIGQUIT);
 	}
 	return ;
 }
 
-void	get_status_philosopher(t_person *philosopher, t_common *common)
+static int	check_status(int status, t_common *common, pid_t temp)
+{
+	if (WIFSIGNALED(status))
+		return (0);
+	if (WIFEXITED(status))
+	{
+		if (WEXITSTATUS(status) != 0)
+		{
+			sem_wait(common->death_sem);
+			kill_others(temp, common->philo_pid, common->init->num_philo);
+			usleep(400);
+			sem_post(common->death_sem);
+		}
+	}
+	return (0);
+}
+
+void	get_status_philosopher(t_common *common)
 {
 	int		i;
 	int		max_len;
@@ -53,19 +67,9 @@ void	get_status_philosopher(t_person *philosopher, t_common *common)
 	while (++i < max_len)
 	{
 		temp = waitpid(-1, &status, 0);
-		if (WIFSIGNALED(status))
-			continue ;
-		if (WIFEXITED(status))
-		{
-			if (WEXITSTATUS(status) != 0)
-			{
-				sem_wait(common->print_sem);
-				printf_action(4, get_ms() - common->dining_time, &philosopher[WEXITSTATUS(status) - 1], common->fork_sem);
-				kill_others(temp, common->philo_pid, max_len);
-				sem_post(common->print_sem);
-			}
-			status = 0;
-		}
+		sem_wait(common->print_sem);
+		status = check_status(status, common, temp);
+		sem_post(common->print_sem);
 	}
 	return ;
 }
